requirements.api.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic-settings
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart
celery[redis]
chromadb-client
langchain
langchain-google-genai
rank-bm25
httpx